+++ example/example4.dart
@@ -1,5 +1,6 @@
 import 'dart:isolate';
 import 'package:arowana/arowana.dart';
+import 'package:arowana/src/file_handle/file_handle.dart';
 import 'package:shelf_web_socket/shelf_web_socket.dart';
 
 base class MyAChannel extends DefaultChannel {
@@ -15,6 +16,9 @@ base class MyAChannel extends DefaultChannel {
     get('/hello', (r) async {
       return Response.ok('hello,arowana! form:${Isolate.current.debugName}');
     });
+
+    // visit http://127.0.0.1:8888/example/example.dart
+    get('/example/*path', createStaticHandler('.'));
   }
 }