+++ lib/src/application/app_channel.dart
@@ -1,31 +1,44 @@
 import 'dart:async';
+import 'package:arowana/src/application/isolate_message.dart';
 import 'package:logging/logging.dart';
 import 'package:shelf/shelf.dart';
 
 import 'application.dart';
 
-abstract class AppChannel{
+abstract base class AppChannel {
+  final ApplicationEventHub eventHub = ApplicationEventHub();
+
   /// The configuration options used to start the application this channel belongs to.
-  ///
-  /// These options are set when starting the application. Changes to this object have no effect
-  /// on other isolates.
-  late ApplicationOptions appOptions;
+  Future<void> initialize(ApplicationOptions options) async {}
 
-  Future initialize(ApplicationOptions options) async{}
-
-  Future prepare()async{}
+  Future<void> prepare() async {}
 
   void entryPoint();
+
+  void emitEvent<T>(T event) {
+    eventHub.add(BroadcastEvent<T>(event));
+  }
+
+  void onEvent<T>(void Function(T data) listener){
+    eventHub.listen((event) {
+      listener(event.payload as T);
+    });
+  }
+
+  Future<void> dispose()async{
+    await eventHub.close();
+  }
 
   FutureOr<Response> call(Request request);
 }
 
-class ApplicationMessageHub extends Stream<dynamic> implements Sink<dynamic> {
+class ApplicationEventHub extends Stream<BroadcastEvent>
+    implements Sink<BroadcastEvent> {
   final Logger _logger = Logger('arowana');
-  final StreamController<dynamic> _outboundController =
-  StreamController<dynamic>();
-  final StreamController<dynamic> _inboundController =
-  StreamController<dynamic>.broadcast();
+  final StreamController<BroadcastEvent> _outboundController =
+      StreamController<BroadcastEvent>();
+  final StreamController<BroadcastEvent> _inboundController =
+      StreamController<BroadcastEvent>.broadcast();
 
   /// Adds a listener for messages from other hubs.
   ///
@@ -35,12 +48,15 @@ class ApplicationMessageHub extends Stream<dynamic> implements Sink<dynamic> {
   /// [onError], if provided, will be invoked when this isolate tries to [add] invalid data. Only the isolate
   /// that failed to send the data will receive [onError] events.
   @override
-  StreamSubscription<dynamic> listen(void Function(dynamic event)? onData,
-      {Function? onError, void Function()? onDone, bool? cancelOnError=false}) =>
+  StreamSubscription<BroadcastEvent> listen(
+          void Function(BroadcastEvent event)? onData,
+          {Function? onError,
+          void Function()? onDone,
+          bool? cancelOnError = false}) =>
       _inboundController.stream.listen(onData,
           onError: onError ??
-                  (err, StackTrace st) =>
-                  _logger.severe('ApplicationMessageHub error', err, st),
+              (err, StackTrace st) =>
+                  _logger.severe('ApplicationEventHub error', err, st),
           onDone: onDone,
           cancelOnError: cancelOnError);
 
@@ -51,12 +67,12 @@ class ApplicationMessageHub extends Stream<dynamic> implements Sink<dynamic> {
   /// [event] must be isolate-safe data - in general, this means it may not be or contain a closure. Consult the API reference `dart:isolate` for more details. If [event]
   /// is not isolate-safe data, an error is delivered to [listen] on this isolate.
   @override
-  void add(dynamic event) {
+  void add(BroadcastEvent event) {
     _outboundController.sink.add(event);
   }
 
   @override
-  Future close() async {
+  Future<void> close() async {
     if (!_outboundController.hasListener) {
       _outboundController.stream.listen(null);
     }