+++ lib/src/application/declaration.dart
@@ -1,0 +1,4 @@
+
+import '../../arowana.dart';
+
+typedef AppChannelBuilder = AppChannel Function();