+++ lib/src/http/request_and_response.dart
@@ -1,12 +1,9 @@
 import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
-
 import 'package:shelf/shelf.dart';
-
-import '../../arowana.dart';
+import '../auth/authorizer.dart';
 import 'body_parse.dart';
-
 
 final _emptyParams = UnmodifiableMapView(<String, String>{});
 
@@ -34,15 +31,14 @@ extension RouterParams on Request {
     return _emptyParams;
   }
 
-  bool get hasQuery =>requestedUri.hasQuery;
+  bool get hasQuery => requestedUri.hasQuery;
 
   Map<String, String> get query {
     return UnmodifiableMapView(url.queryParameters);
   }
 }
 
-extension BodyParams on Request{
-
+extension BodyParams on Request {
   bool get hasFormData {
     var contentTypeStr = headers[HttpHeaders.contentTypeHeader];
     if (contentTypeStr != null) {
@@ -53,9 +49,9 @@ extension BodyParams on Request{
     return false;
   }
 
-  Future<RequestBody> get body async{
+  Future<RequestBody> get body async {
     var rb = context['arowana/RequestBody'] as RequestBody?;
-    if(rb == null){
+    if (rb == null) {
       rb = RequestBody();
       await rb.parseBody(this);
       change(context: {'arowana/RequestBody': rb});
@@ -64,37 +60,17 @@ extension BodyParams on Request{
   }
 }
 
-
-
-extension ResponseX on Response{
-
+extension ResponseX on Response {
   /// Constructs a 200 OK response.
-  static Response ok(Map body,{Map<String,Object>? headers}){
-    return Response.ok(json.encode(body),headers: headers);
-  }
-
-  /// Represents a 400 response.
-  static Response badRequest(body,{Map<String,Object>? headers}){
-    return Response(HttpStatus.badRequest,body: body,headers: headers);
-  }
-
-  /// Represents a 401 response.
-  static Response unauthorized(body,{Map<String,Object>? headers}){
-    return Response(HttpStatus.unauthorized,body: body);
-  }
-
-  /// Represents a 500 response.
-  static Response serverError(body,{Map<String,Object>? headers}){
-    return Response(HttpStatus.internalServerError,body: body);
+  static Response ok(Map body, {Map<String, Object>? headers}) {
+    return Response.ok(json.encode(body), headers: headers);
   }
 
   static Response token(AuthToken token) {
-    return Response(HttpStatus.ok,
-        headers: {
-          'Cache-Control': 'no-store',
-          'Pragma': 'no-cache',
-          HttpHeaders.contentTypeHeader: 'application/json'
-        },
-        body: json.encode(token.asMap()));
+    return Response.ok(json.encode(token.asMap()), headers: {
+      'Cache-Control': 'no-store',
+      'Pragma': 'no-cache',
+      HttpHeaders.contentTypeHeader: 'application/json'
+    });
   }
 }