+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+- Upgrade to dart 3.x support
+- Added a websocket handler that can be used to provide websocket services
+- Added a static file serving handler to serve files
+- Fixed the error reporting for dart 3 isolate concurrency support
+
 ## 0.0.2
 
 - Fix dart version 3 error reporting